Make VTT generation faster and more memory/cache efficient
The VTT generation process was generating a 16.5mb image for
each caption.  We now cache a single image and draw text all
over it since the end user will never see that image.

This was originally so that we could overlay the tmp images
over our videos to check for alignment and how closely
they resemble the captions we created and positioned manually.

We no longer need that level of detail.